# backend/docs/swagger.yaml
definitions:
...
  dto.AuthResponse:
    properties:
      access_expires_at:
        type: string
      access_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  dto.ProfileResponse:
    properties:
      permissions:
        items:
          type: string
        type: array
      roles:
        items:
          $ref: '#/definitions/dto.RoleResponse'
        type: array
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.RegisterRequest:
    properties:
      birthday:
        type: string
      email:
        type: string
      first_name:
        type: string
      group_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      password:
        type: string
    type: object
...
  dto.RoleResponse:
    properties:
      id:
        type: string
      level:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      birthday:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      group_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email и пароль
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Неверные учётные данные
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Вход в систему
      tags:
      - auth
  /api/auth/logout:
    post:
      responses:
        "204":
          description: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Выход из системы
      tags:
      - auth
  /api/auth/me:
    get:
      description: Возвращает пользователя, его роли и плоский список permissions.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Профиль текущего пользователя
      tags:
      - auth
  /api/auth/refresh:
    post:
      description: 'Читает refresh-токен из httpOnly-cookie, возвращает новую пару.
        One-use: старый токен инвалидируется.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "401":
          description: Cookie отсутствует или токен недействителен
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Обновление access-токена
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для регистрации
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Email уже занят
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - auth
...
swagger: "2.0"